Coding Team/LMS Agent/New Lms Agent/requirements.txt
fastapi
pydantic
uvicorn
langchain
google-generativeai
pymongo
faiss-cpu
python-dotenv
jinja2
